refactor: WorkoutCalendarHistoryViewのリファクタリング
- WorkoutCalendarHistoryView.swift（373行）を5ファイルに分割
- Views/WorkoutCalendarHistory/ フォルダを作成

分割結果:
- WorkoutCalendarHistoryView.swift (69行) - メインビュー
- CalendarHistoryHeaderView.swift (79行) - ヘッダー・曜日
- CalendarHistoryGridView.swift (74行) - カレンダーグリッド
- CalendarHistoryDateCell.swift (79行) - 日付セル
- SelectedWorkoutDetailView.swift (66行) - ワークアウト詳細

Closes #117
